+++ app/src/main/java/com/example/pomodoro/model/PomodoroMode.kt
@@ -1,4 +1,4 @@
-package com.example.pomodoro
+package com.example.pomodoro.model
 
 import androidx.compose.ui.graphics.Color
