+++ app/src/main/java/com/example/pomodoro/ui/screens/HistoryScreen.kt
@@ -1,0 +1,10 @@
+package com.example.pomodoro.ui.screens
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun HistoryScreen() {
+    Text("Histórico")
+
+}
